requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-xml==1.4.0
entrypoints==0.3
gunicorn==20.0.4
Markdown==3.2.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
